pkglistgen: Introduce file_utils functions
